+++ src/ts/util/Options.ts
@@ -30,6 +30,7 @@ export class Options {
         preview: {
             delay: 1000,
             hljs: {
+                lineNumber: false,
                 enable: true,
                 style: "github",
             },
@@ -220,13 +221,10 @@ export class Options {
             }
 
             if (this.options.preview) {
-                if (this.options.preview.hljs) {
-                    this.options.preview.hljs =
-                        Object.assign({}, this.defaultOptions.preview.hljs, this.options.preview.hljs);
-                } else {
-                    this.options.preview.hljs = Object.assign({}, this.defaultOptions.preview.hljs);
-                }
                 this.options.preview = Object.assign({}, this.defaultOptions.preview, this.options.preview);
+                if (this.options.preview.hljs) {
+                    this.options.preview.hljs = Object.assign({}, this.defaultOptions.preview.hljs, this.options.preview.hljs);
+                }
             }
 
             if (this.options.hint) {